requirement.txt
fastapi==0.68.1
uvicorn
python-socketio[asyncio_client]==5.4.1
psycopg2-binary
python-dotenv==0.19.1
boto3==1.18.61
sqlalchemy==1.4.20
pydantic==1.8.2
bcrypt==3.2.0
oauth2==1.9.0.post1
uvicorn==0.14.0
gunicorn==20.1.0
aioredis==2.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-jose